fix: remove search query and results from local storage and retrieve data using search params
