+++ pages/marketplace/search/[query].tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import CategoriesNav from '@modules/marketplace/component/CategoriesNav/CategoriesNav';
-import MainLayout from '../../../components/Layout/MainLayout';
 import { useEffect, useState } from 'react';
 import ProductCard from '@modules/marketplace/component/ProductCard';
 import styles from '../../../modules/marketplace/component/landingpage/productCardWrapper/product-card-wrapper.module.css';
@@ -9,30 +7,40 @@ import { ProductResult } from '../../../@types';
 import Link from 'next/link';
 import Error from '@modules/marketplace/component/landingpageerror/ErrorPage';
 import CategoryLayout from '@modules/marketplace/component/layout/category-layout';
+import { useRouter } from 'next/router';
+import { searchPosts } from '../../../http/api/searchPost';
 
 export default function Index() {
   const [results, setResults] = useState<ProductResult[]>([]);
-  const searchquery = typeof window !== 'undefined' ? localStorage.getItem('keyword') : null;
+  const router = useRouter();
+  const { query } = router.query;
+  const searchQuery = Array.isArray(query) ? query[0] : query;
 
   useEffect(() => {
-    if (typeof window !== 'undefined') {
-      const getresult = localStorage.getItem('search_result');
-      if (getresult) {
-        const result = JSON.parse(getresult);
-        setResults(result);
-      }
+    if (searchQuery) {
+      const fetchData = async () => {
+        try {
+          const results = await searchPosts(searchQuery);
+          setResults(results);
+          console.log(results)
+        } catch (error) {
+          console.error(error);
+        }
+      };
+
+      fetchData();
     }
-  }, [results]);
+  }, [searchQuery]);
 
   return (
     <>
-      {results?.length === 0 ? (
+    {results?.length === 0 ? (
         <Error />
       ) : (
         <CategoryLayout>
           <div className="px-4 py-4 sm:py-2 max-w-[1240px] mx-auto">
             <h1 className="text-custom-color31 font-manropeL mt-5 lg:pt-5 md:mb-1 font-bold md:text-2xl leading-normal flex items-center justify-between">
-              Search Result for &apos;{searchquery}&apos;
+              Search Result for &apos;{searchQuery}&apos;
             </h1>
             <div
               className={`flex py-8 flex-nowrap lg:flex-wrap gap-y-[70px] mb-[74px] w-full overflow-scroll ${styles['hide-scroll']}`}
@@ -70,3 +78,4 @@ export default function Index() {
     </>
   );
 }
+
